# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TASK MANAGER API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/add/image": {
            "post": {
                "tags": [
                    "User API"
                ],
                "summary": "Register user",
                "operationId": "4418c8a0bd16417e782b4086aec70325",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "type": "string",
                                        "example": "upload/image/file/rHt7PfnEqNS8cpXBOkOn.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registration user and to login.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": {
                                                "username": "Abdullajon",
                                                "password": "$2y$10$sZZUV7ykVyAXT7/eDzGyReThtdkaLUK9jknrDEDNaeK71uKeQkfiO",
                                                "token": "ms7ST2opzINdvWCxLanr8zg21kOKFS",
                                                "updated_at": "2023-08-28T11:58:12.000000Z",
                                                "created_at": "2023-08-28T11:58:12.000000Z",
                                                "id": 1
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Login!"
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}